// base/interfaces.ts
export type IDocumentReadyState = 'complete' | 'interactive' | 'loading';
// ...
export type IVisibilityState = 'hidden' | 'prerender' | 'visible';

export type IEventListenerOrEventListenerObject = IEventListener | IEventListenerObject;

export type IUnsupportedFeature = void;

export type ICanvasRenderingContext2D = IUnsupportedFeature;

export type IWebGLRenderingContext = IUnsupportedFeature;

export type IWindowProxy = IUnsupportedFeature;
// ...
export interface IElementCreationOptions {
  is?: string;
}
// ...
export interface ICharacterData extends INode, IChildNode, INonDocumentTypeChildNode {
  data: string;
  readonly length: number;
  appendData(data: string): void;
  deleteData(offset: number, count: number): void;
  insertData(offset: number, data: string): void;
  replaceData(offset: number, count: number, data: string): void;
  substringData(offset: number, count: number): string;
}

export interface IChildNode {
  after(...nodes: (INode | string)[]): void;
  before(...nodes: (INode | string)[]): void;
  remove(): void;
  replaceWith(...nodes: (INode | string)[]): void;
}

export interface IComment extends ICharacterData {
}
// ...
export interface IDocument extends INode, IDocumentOrShadowRoot, IGlobalEventHandlers, INonElementParentNode, IParentNode {
  readonly URL: string;
  readonly anchors: IHTMLCollection;
  readonly applets: IHTMLCollection;
  body: IHTMLElement | null;
  readonly characterSet: string;
  readonly charset: string;
  readonly compatMode: string;
  readonly contentType: string;
  cookie: string;
  readonly currentScript: IHTMLScriptElement | null;
  readonly defaultView: IWindowProxy | null;
  dir: string;
  readonly doctype: IDocumentType | null;
  readonly documentElement: IElement | null;
  readonly documentURI: string;
  readonly embeds: IHTMLCollection;
  readonly forms: IHTMLCollection;
  readonly head: IHTMLHeadElement | null;
  readonly hidden: boolean;
  readonly images: IHTMLCollection;
  readonly implementation: IDOMImplementation;
  readonly inputEncoding: string;
  readonly lastModified: string;
  readonly links: IHTMLCollection;
  readonly location: ILocation | null;
  onreadystatechange: ((this: IDocument, ev: IEvent) => any) | null;
  onvisibilitychange: ((this: IDocument, ev: IEvent) => any) | null;
  readonly plugins: IHTMLCollection;
  readonly readyState: IDocumentReadyState;
  readonly referrer: string;
  readonly scripts: IHTMLCollection;
  title: string;
  readonly visibilityState: IVisibilityState;
  adoptNode(node: INode): INode;
  captureEvents(): void;
  clear(): void;
  close(): void;
  createAttribute(localName: string): IAttr;
  createAttributeNS(namespace: string | null, qualifiedName: string): IAttr;
  createCDATASection(data: string): ICDATASection;
  createComment(data: string): IComment;
  createDocumentFragment(): IDocumentFragment;
  createElement(localName: string, options?: IElementCreationOptions): IElement;
  createElementNS(namespace: string | null, qualifiedName: string, options?: IElementCreationOptions): IElement;
  createEvent(eventInterface: string): IEvent;
  createNodeIterator(root: INode, whatToShow?: number, filter?: INodeFilter | null): INodeIterator;
  createProcessingInstruction(target: string, data: string): IProcessingInstruction;
  createRange(): IRange;
  createTextNode(data: string): IText;
  createTreeWalker(root: INode, whatToShow?: number, filter?: INodeFilter | null): ITreeWalker;
  getElementsByClassName(classNames: string): IHTMLCollection;
  getElementsByName(elementName: string): INodeList;
  getElementsByTagName(qualifiedName: string): IHTMLCollection;
  getElementsByTagNameNS(namespace: string | null, localName: string): IHTMLCollection;
  getSelection(): ISelection | null;
  hasFocus(): boolean;
  importNode(node: INode, deep?: boolean): INode;
  open(type?: string, replace?: string): IDocument;
  releaseEvents(): void;
  write(...text: string[]): void;
  writeln(...text: string[]): void;
  addEventListener<K extends keyof IDocumentEventMap>(type: K, listener: (this: IDocument, ev: IDocumentEventMap[K]) => any, options?: boolean | IAddEventListenerOptions): void;
  addEventListener(type: string, listener: IEventListenerOrEventListenerObject, options?: boolean | IAddEventListenerOptions): void;
  removeEventListener<K extends keyof IDocumentEventMap>(type: K, listener: (this: IDocument, ev: IDocumentEventMap[K]) => any, options?: boolean | IEventListenerOptions): void;
  removeEventListener(type: string, listener: IEventListenerOrEventListenerObject, options?: boolean | IEventListenerOptions): void;
}

export interface IDocumentFragment extends INode, INonElementParentNode, IParentNode {
}

export interface IDocumentOrShadowRoot {
  readonly activeElement: IElement | null;
}

export interface IDocumentType extends INode, IChildNode {
  readonly name: string;
  readonly publicId: string;
  readonly systemId: string;
}

export interface IElement extends INode, IChildNode, INonDocumentTypeChildNode, IParentNode, ISlotable {
  readonly attributes: INamedNodeMap;
  readonly classList: IDOMTokenList;
  className: string;
  readonly clientHeight: number;
  readonly clientLeft: number;
  readonly clientTop: number;
  readonly clientWidth: number;
  id: string;
  innerHTML: string;
  readonly localName: string;
  readonly namespaceURI: string | null;
  outerHTML: string;
  readonly prefix: string | null;
  readonly scrollHeight: number;
  scrollLeft: number;
  scrollTop: number;
  readonly scrollWidth: number;
  readonly shadowRoot: IShadowRoot | null;
  slot: string;
  readonly tagName: string;
  attachShadow(init: IShadowRootInit): IShadowRoot;
  closest(selectors: string): IElement | null;
  getAttribute(qualifiedName: string): string | null;
  getAttributeNS(namespace: string | null, localName: string): string | null;
  getAttributeNames(): Iterable<string>;
  getAttributeNode(qualifiedName: string): IAttr | null;
  getAttributeNodeNS(namespace: string | null, localName: string): IAttr | null;
  getBoundingClientRect(): IDOMRect;
  getClientRects(): IDOMRectList;
  getElementsByClassName(classNames: string): IHTMLCollection;
  getElementsByTagName(qualifiedName: string): IHTMLCollection;
  getElementsByTagNameNS(namespace: string | null, localName: string): IHTMLCollection;
  hasAttribute(qualifiedName: string): boolean;
  hasAttributeNS(namespace: string | null, localName: string): boolean;
  hasAttributes(): boolean;
  insertAdjacentElement(where: string, element: IElement): IElement | null;
  insertAdjacentHTML(position: string, text: string): void;
  insertAdjacentText(where: string, data: string): void;
  matches(selectors: string): boolean;
  removeAttribute(qualifiedName: string): void;
  removeAttributeNS(namespace: string | null, localName: string): void;
  removeAttributeNode(attr: IAttr): IAttr;
  setAttribute(qualifiedName: string, value: string): void;
  setAttributeNS(namespace: string | null, qualifiedName: string, value: string): void;
  setAttributeNode(attr: IAttr): IAttr | null;
  setAttributeNodeNS(attr: IAttr): IAttr | null;
  toggleAttribute(qualifiedName: string, force?: boolean): boolean;
  webkitMatchesSelector(selectors: string): boolean;
}
// ...
export interface IHTMLCollection<T extends IElement = IElement> extends IHTMLCollectionBase {
  namedItem(name: string): T | null;
}
// ...
export interface INonElementParentNode {
  getElementById(elementId: string): IElement | null;
}

export interface IParentNode {
  readonly childElementCount: number;
  readonly children: IHTMLCollection;
  readonly firstElementChild: IElement | null;
  readonly lastElementChild: IElement | null;
  append(...nodes: (INode | string)[]): void;
  prepend(...nodes: (INode | string)[]): void;
  querySelector(selectors: string): IElement | null;
  querySelectorAll(selectors: string): INodeList;
}
// ...
export interface IText extends ICharacterData, ISlotable {
  readonly wholeText: string;
  splitText(offset: number): IText;
}
// ...
export interface IHTMLElementTagNameMap {
  a: IHTMLAnchorElement;
  abbr: IHTMLElement;
  acronym: IHTMLElement;
  address: IHTMLElement;
  area: IHTMLAreaElement;
  article: IHTMLElement;
  aside: IHTMLElement;
  audio: IHTMLAudioElement;
  b: IHTMLElement;
  base: IHTMLBaseElement;
  bdi: IHTMLElement;
  bdo: IHTMLElement;
  big: IHTMLElement;
  blockquote: IHTMLQuoteElement;
  body: IHTMLBodyElement;
  br: IHTMLBRElement;
  button: IHTMLButtonElement;
  canvas: IHTMLCanvasElement;
  caption: IHTMLTableCaptionElement;
  center: IHTMLElement;
  cite: IHTMLElement;
  code: IHTMLElement;
  col: IHTMLTableColElement;
  colgroup: IHTMLTableColElement;
  data: IHTMLDataElement;
  datalist: IHTMLDataListElement;
  dd: IHTMLElement;
  del: IHTMLModElement;
  details: IHTMLDetailsElement;
  dfn: IHTMLElement;
  dialog: IHTMLDialogElement;
  dir: IHTMLDirectoryElement;
  div: IHTMLDivElement;
  dl: IHTMLDListElement;
  dt: IHTMLElement;
  em: IHTMLElement;
  embed: IHTMLEmbedElement;
  fieldset: IHTMLFieldSetElement;
  figcaption: IHTMLElement;
  figure: IHTMLElement;
  font: IHTMLFontElement;
  footer: IHTMLElement;
  form: IHTMLFormElement;
  frame: IHTMLFrameElement;
  frameset: IHTMLFrameSetElement;
  h1: IHTMLHeadingElement;
  h2: IHTMLHeadingElement;
  h3: IHTMLHeadingElement;
  h4: IHTMLHeadingElement;
  h5: IHTMLHeadingElement;
  h6: IHTMLHeadingElement;
  head: IHTMLHeadElement;
  header: IHTMLElement;
  hgroup: IHTMLElement;
  hr: IHTMLHRElement;
  html: IHTMLHtmlElement;
  i: IHTMLElement;
  iframe: IHTMLIFrameElement;
  img: IHTMLImageElement;
  input: IHTMLInputElement;
  ins: IHTMLModElement;
  isindex: IHTMLUnknownElement;
  kbd: IHTMLElement;
  keygen: IHTMLElement;
  label: IHTMLLabelElement;
  legend: IHTMLLegendElement;
  li: IHTMLLIElement;
  link: IHTMLLinkElement;
  main: IHTMLElement;
  map: IHTMLMapElement;
  mark: IHTMLElement;
  marquee: IHTMLMarqueeElement;
  menu: IHTMLMenuElement;
  meta: IHTMLMetaElement;
  meter: IHTMLMeterElement;
  nav: IHTMLElement;
  nextid: IHTMLUnknownElement;
  nobr: IHTMLElement;
  noframes: IHTMLElement;
  noscript: IHTMLElement;
  object: IHTMLObjectElement;
  ol: IHTMLOListElement;
  optgroup: IHTMLOptGroupElement;
  option: IHTMLOptionElement;
  output: IHTMLOutputElement;
  p: IHTMLParagraphElement;
  param: IHTMLParamElement;
  picture: IHTMLPictureElement;
  plaintext: IHTMLElement;
  pre: IHTMLPreElement;
  progress: IHTMLProgressElement;
  q: IHTMLQuoteElement;
  rp: IHTMLElement;
  rt: IHTMLElement;
  ruby: IHTMLElement;
  s: IHTMLElement;
  samp: IHTMLElement;
  script: IHTMLScriptElement;
  section: IHTMLElement;
  select: IHTMLSelectElement;
  slot: IHTMLSlotElement;
  small: IHTMLElement;
  source: IHTMLSourceElement;
  span: IHTMLSpanElement;
  strike: IHTMLElement;
  strong: IHTMLElement;
  style: IHTMLStyleElement;
  sub: IHTMLElement;
  summary: IHTMLElement;
  sup: IHTMLElement;
  table: IHTMLTableElement;
  tbody: IHTMLTableSectionElement;
  td: IHTMLTableCellElement;
  template: IHTMLTemplateElement;
  textarea: IHTMLTextAreaElement;
  tfoot: IHTMLTableSectionElement;
  th: IHTMLTableCellElement;
  thead: IHTMLTableSectionElement;
  time: IHTMLTimeElement;
  title: IHTMLTitleElement;
  tr: IHTMLTableRowElement;
  track: IHTMLTrackElement;
  tt: IHTMLElement;
  u: IHTMLElement;
  ul: IHTMLUListElement;
  var: IHTMLElement;
  video: IHTMLVideoElement;
  wbr: IHTMLElement;
}
// ...
export interface IHTMLElement extends IElement, IElementCSSInlineStyle, IElementContentEditable, IGlobalEventHandlers, IHTMLOrSVGElement {
  accessKey: string;
  dir: string;
  draggable: boolean;
  hidden: boolean;
  lang: string;
  readonly offsetHeight: number;
  readonly offsetLeft: number;
  readonly offsetParent: IElement | null;
  readonly offsetTop: number;
  readonly offsetWidth: number;
  title: string;
  translate: boolean;
  click(): void;
  addEventListener<K extends keyof IGlobalEventHandlersEventMap>(type: K, listener: (this: IHTMLElement, ev: IGlobalEventHandlersEventMap[K]) => any, options?: boolean | IAddEventListenerOptions): void;
  addEventListener(type: string, listener: IEventListenerOrEventListenerObject, options?: boolean | IAddEventListenerOptions): void;
  removeEventListener<K extends keyof IGlobalEventHandlersEventMap>(type: K, listener: (this: IHTMLElement, ev: IGlobalEventHandlersEventMap[K]) => any, options?: boolean | IEventListenerOptions): void;
  removeEventListener(type: string, listener: IEventListenerOrEventListenerObject, options?: boolean | IEventListenerOptions): void;
}